test: split core.multiplexer tests into separate files
